Pages/requirements.txt
numpy==1.23.2
pandas==1.4.4
Pillow==9.2.0
plotly==5.10.0
plotly_express==0.4.1
streamlit==1.12.2
streamlit-aggrid==0.3.3
streamlit-option-menu==0.3.2
